moved NewUUID and Config to separate files
