+++ extension/goext/util.go
@@ -1,5 +1,6 @@
 package goext
 
 type IUtil interface {
+	NewUUID() string
 	GetTransaction(context Context) (ITransaction, bool)
 }
