+++ extension/goplugin/util.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/cloudwan/gohan/db/transaction"
 	"github.com/cloudwan/gohan/extension/goext"
+	"github.com/twinj/uuid"
 )
 
 type Util struct {
@@ -24,6 +25,11 @@ func contextGetTransaction(ctx goext.Context) (goext.ITransaction, bool) {
 	default:
 		panic(fmt.Sprintf("Unknown transaction type in context: %+v", ctxTx))
 	}
+}
+
+// NewUUID create a new unique ID
+func (util *Util) NewUUID() string {
+	return uuid.NewV4().String()
 }
 
 func (u *Util) GetTransaction(context goext.Context) (goext.ITransaction, bool) {
